+++ language/en/install/database.php
@@ -1,0 +1,4 @@
+<?php
+
+$lang['ins_database_close_reason'] = 'Sorry. The game is under maintenance at the moment.';
+$lang['ins_database_news'] = 'Welcome to the new XNova server!';