+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ### Changed
 - Improved translation of
     - Log in screen
+- Installer improvements
+    - Default database texts (like overview news) are now localized
 - Signin no longer gives a hint whether username or password was incorrect
 
 ### Removed